% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertDataFrameListOfVectors.R
\name{convertDataFrameListOfVectors}
\alias{convertDataFrameListOfVectors}
\title{Convert (simplify) data.frames with list of vectors to single entries}
\usage{
convertDataFrameListOfVectors(df, delimiter = "|")
}
\arguments{
\item{df}{a data.frame}

\item{delimiter}{a delimiter between list items; default: | (vertical bar)}
}
\value{
a new data.frame
}
\description{
Convert (simplify) data.frames with list of vectors to single entries
}
\examples{
b <- list() 
b[[1]] <- c("a","b","c")
b[[2]] <- c("d","e","f")
b[[3]] <- c("g","h","i")
df <- data.frame(a=c(1,2,3), b=I(b), stringsAsFactors=FALSE)
convertDataFrameListOfVectors(df, ";")
  
}
\concept{
listutils
}
