% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertDataFrameToList.R
\name{convertDataFrameToList}
\alias{convertDataFrameToList}
\title{Convert a data.frame to a list}
\usage{
convertDataFrameToList(df, listItemNameCol, listItemCol)
}
\arguments{
\item{df}{a data.frame}

\item{listItemNameCol}{the name of the column that contains the categories to be split on}

\item{listItemCol}{the name of the column that contains the entries to be added to list entries}
}
\description{
Convert a data.frame to a list
}
\examples{
"TODO"
  
}
\concept{
listutils
}
