% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertListOfVectorsDataFrameToDataFrame.R
\name{convertListOfVectorsDataFrameToDataFrame}
\alias{convertListOfVectorsDataFrameToDataFrame}
\title{Convert a data.frame containing lists of vectors to a simple data.frame}
\usage{
convertListOfVectorsDataFrameToDataFrame(x, sep2)
}
\arguments{
\item{x}{a data.frame}

\item{sep2}{a separator to use between vector entries}
}
\value{
a data.frame
}
\description{
Convert a data.frame containing lists of vectors to a simple data.frame
}
\concept{
listutils
}
