% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertListOfVectorsToDataFrame.R
\name{convertListOfVectorsToDataFrame}
\alias{convertListOfVectorsToDataFrame}
\title{Convert List of Vectors to data.frame}
\usage{
convertListOfVectorsToDataFrame(lst, colNames = NULL,
  collapseDelimiter = NULL)
}
\arguments{
\item{lst}{a list}

\item{colNames}{an optional vector of column names for the resulting data.frame}

\item{collapseDelimiter}{a delimiter to collapse the vector; optional}
}
\value{
a data.frame
}
\description{
Convert List of Vectors to data.frame
}
\examples{
lov <- list(v1=c("a", "b"), v2=c("a", "c"), v3=c("a"))
convertListOfVectorsToDataFrame(lov)

}
\concept{
listutils
}
