% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToDataFrameWithListOfVectors.R
\name{convertToDataFrameWithListOfVectors}
\alias{convertToDataFrameWithListOfVectors}
\title{Convert data.frame to a list of vectors}
\usage{
convertToDataFrameWithListOfVectors(df, delimiter = ";")
}
\arguments{
\item{df}{a data.frame where some entries may be delimited}

\item{delimiter}{a delimiter between entries; default: ";"}
}
\value{
a list entries converted to data.frame with lists of vectors
}
\description{
Convert data.frame to a list of vectors
}
\examples{
"TODO"
  
}
\concept{
listutils
}
