% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverseListOfVectors.R
\name{inverseListOfVectors}
\alias{inverseListOfVectors}
\title{Inverse a list of vectors}
\usage{
inverseListOfVectors(lov)
}
\arguments{
\item{lov}{a list of vectors}
}
\value{
the inverted list
}
\description{
Inverse a list of vectors
}
\details{
From: https://stackoverflow.com/questions/35827008/invert-a-list-of-character-vectors-in-r
}
