% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchListOfVectors.R
\name{searchListOfVectors}
\alias{searchListOfVectors}
\title{Search List of Vectors}
\usage{
searchListOfVectors(q, lst, useNames = FALSE, queryAll = FALSE)
}
\arguments{
\item{q}{query vector}

\item{lst}{list of vectors to search}

\item{useNames}{return names rather than indicies; ignored if no names. Default: FALSE}

\item{queryAll}{query for union of q entries}
}
\value{
a list of vectors with the same length as the query vector, each list
  entry will have indicies for lst where there was a match with the query
  vector. Return NA if there were no matches.
}
\description{
Search List of Vectors
}
\details{
Taken from: http://stackoverflow.com/questions/11002391/fast-way-of-getting-index-of-match-in-list
}
\examples{
lst <- list(1:3, 3:5, 3:7)
q <- c(3, 5)
results <- searchListOfVectors(q, lst)
names(results) <- q

lst <- list(LETTERS[1:3], LETTERS[3:5], LETTERS[3:7])
q <- c("C", "E")
searchListOfVectors(q, lst)

lst <- list(LETTERS[3], LETTERS[4:6])
q <- "C"
searchListOfVectors(q, lst)

lst <- list(LETTERS[3], LETTERS[4:6])
q <- c("C")
searchListOfVectors(q, lst)

lst <- list(LETTERS[3], LETTERS[4:6])
q <- c("C", "E")
searchListOfVectors(q, lst)

lst <- list(a=LETTERS[3], b=LETTERS[4:6])
q <- c("C", "E")
searchListOfVectors(q, lst, useNames=TRUE)

lst <- list(a=LETTERS[3], b=LETTERS[3:6])
q <- c("C", "E")
searchListOfVectors(q, lst, useNames=TRUE)

lst <- list(LETTERS[3], LETTERS[4:6])
q <- "Z"
searchListOfVectors(q, lst)

lst <- list(a=LETTERS[3], b=LETTERS[3:6])
q <- c("C", "E")
searchListOfVectors(q, lst, useNames=TRUE, queryAll=TRUE)

lst <- list(a=LETTERS[3], b=LETTERS[3:6])
q <- c("C", "E")
searchListOfVectors(q, lst, useNames=FALSE, queryAll=TRUE)

#lst <- list(LETTERS[3], LETTERS[3:6])
#q <- c("C", "E")
#searchListOfVectors(q, lst, useNames=TRUE, queryAll=TRUE)
  
}
\concept{
listutils
}
